+++ lib/plug/accesslog/application.ex
@@ -7,11 +7,19 @@ defmodule Plug.AccessLog.Application do
 
   use Application
 
+  alias Plug.AccessLog.Logfiles
+  alias Plug.AccessLog.Writer
+
+
   def start(_type, _args) do
     import Supervisor.Spec
 
     options  = [ strategy: :one_for_one, name: __MODULE__.Supervisor ]
-    children = [ worker(Plug.AccessLog.Logfiles, []) ]
+    children = [
+      worker(Logfiles, []),
+      worker(GenEvent, [[ name: Writer ]]),
+      worker(Writer.Watcher, [ Writer ])
+    ]
 
     Supervisor.start_link(children, options)
   end